requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
openai
elevenlabs
deepgram-sdk
cartesia
pydantic
httpx
websockets
python-multipart
